requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==22.1.0
awkward==1.10.2
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cffi==1.15.1
configparser==5.3.0
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==1.0.0
fastjsonschema==2.16.1
fonttools==4.37.1
ipykernel==6.15.2
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.15.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter_client==7.3.5
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.2
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mistune==2.0.4
nbclient==0.6.7
nbconvert==7.0.0
nbformat==5.4.0
nest-asyncio==1.5.5
notebook==6.4.12
numpy==1.23.2
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
pdf2image==1.16.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
PyPDF2==2.11.0
PyQt5==5.15.7
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.0
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.2.1
pyzmq==23.2.1
qtconsole==5.3.2
QtPy==2.2.0
reportlab==3.6.11
scikit-learn==1.1.3
scipy==1.9.1
seaborn==0.11.2
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.2.post1
stack-data==0.5.0
terminado==0.15.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tk==0.1.0
tornado==6.2
traitlets==5.3.0
uproot==4.3.4
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==4.0.2
